fix(roles): relocate role roster to /Profile/Admin/Roles + dynamic filter
The role-assignment roster lived at /Governance/Roles, but role_assignments
is owned by Auth, not Governance (docs/sections/Auth.md), and its per-human
management (AssignRole/EndRole, /Profile/{id}/Admin/Roles/*) already lives in
the Profiles section. Move the system-wide roster beside it at
/Profile/Admin/Roles and re-home the admin-nav entry from the Governance
group to Members. Authorization stays BoardOrAdmin (unchanged).

The filter pill bar was hardcoded to All/Admin/Board. Drive it from a new
RoleNames.All ordered catalog so every defined role surfaces automatically;
RoleNamesTests fails if a future role constant is omitted from All.
