+++ src/Humans.Web/Views/Shared/_RolesListContent.cshtml
@@ -7,19 +7,18 @@
 
 <div class="card mb-4">
     <div class="card-body d-flex justify-content-between align-items-center">
-        <div class="btn-group" role="group">
+        <div class="d-flex flex-wrap gap-2">
             <a asp-action="Roles" asp-route-showInactive="@Model.ShowInactive"
                class="btn @(string.IsNullOrEmpty(Model.RoleFilter) ? "btn-primary" : "btn-outline-primary")">
                 @Localizer["AdminRoles_All"]
             </a>
-            <a asp-action="Roles" asp-route-role="Admin" asp-route-showInactive="@Model.ShowInactive"
-               class="btn @(Model.RoleFilter == "Admin" ? "btn-primary" : "btn-outline-primary")">
-                Admin
-            </a>
-            <a asp-action="Roles" asp-route-role="Board" asp-route-showInactive="@Model.ShowInactive"
-               class="btn @(Model.RoleFilter == "Board" ? "btn-primary" : "btn-outline-primary")">
-                Board
-            </a>
+            @foreach (var role in Humans.Domain.Constants.RoleNames.All)
+            {
+                <a asp-action="Roles" asp-route-role="@role" asp-route-showInactive="@Model.ShowInactive"
+                   class="btn @(Model.RoleFilter == role ? "btn-primary" : "btn-outline-primary")">
+                    @role
+                </a>
+            }
         </div>
         <div>
             @if (Model.ShowInactive)
